requirements.txt
streamlit
joblib
scikit-learn
pandas
numpy
requests
psycopg2-binary